move cumulative distributed calculations to dedicated class
